+++ app/policies/announcement_policy.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AnnouncementPolicy < ApplicationPolicy
+  class Scope < ApplicationPolicy::Scope
+    def resolve
+      scope.all
+    end
+  end
+end